+++ src/composables/usePointerAction.js
@@ -1,4 +1,5 @@
 import { toRefs, watch, nextTick, computed } from 'vue'
+import toRef from './../utils/toRef'
 
 export default function usePointer (props, context, dep)
 {
@@ -32,11 +33,11 @@ export default function usePointer (props, context, dep)
     return fg.value.filter(g => !g[disabledProp.value])
   })
 
-  const canPointGroups = computed(() => {
+  const canPointGroups = toRef(() => {
     return mode.value !== 'single' && groupSelect.value
   })
 
-  const isPointerGroup = computed(() => {
+  const isPointerGroup = toRef(() => {
     return pointer.value && pointer.value.group
   })
 
@@ -240,7 +241,7 @@ export default function usePointer (props, context, dep)
   })
 
   watch(isOpen, (val) => {
-    if (val && multiselect?.value) {
+    if (val && multiselect && multiselect.value) {
       let firstSelected = multiselect.value.querySelectorAll(`[data-selected]`)[0]
 
       if (!firstSelected) {
@@ -250,10 +251,11 @@ export default function usePointer (props, context, dep)
       let wrapper = firstSelected.parentElement.parentElement
       
       nextTick(() => {
+        // Removed because of #406
         /* istanbul ignore next */
-        if (wrapper.scrollTop > 0) {
-          return
-        }
+        // if (wrapper.scrollTop > 0) {
+        //   return
+        // }
 
         wrapper.scrollTop = firstSelected.offsetTop
       })