+++ tests/unit/composables/useOptions.spec.js
@@ -30,6 +30,31 @@ describe('useOptions', () => {
     })
   })
 
+  describe('resolvedOptions', () => {
+    it('should set and get `resolvedOptions`', () => {
+      let select = createSelect({
+        value: null,
+        options: [1,2],
+      })
+
+      expect(select.vm.resolvedOptions).toEqual([1,2])
+
+      select.vm.resolvedOptions = [1,2,3]
+      
+      expect(select.vm.resolvedOptions).toEqual([1,2,3])
+    })
+    it('should be limit if infinite=true & limit=20', () => {
+      let select = createSelect({
+        value: null,
+        options: [1,2,3],
+        infinite: true,
+        limit: 20,
+      })
+
+      expect(select.vm.offset).toStrictEqual(20)
+    })
+  })
+
   describe('fo', () => {
     it('should be an empty array of options not defined', () => {
       let select = createSelect()
@@ -1640,6 +1665,22 @@ describe('useOptions', () => {
 
       expect(select.vm.isSelected({ v: 2, label: 2 })).toBe(true)
       expect(select.vm.isSelected({ v: 3, label: 3 })).toBe(false)
+    })
+
+    it('should be true if the valueProp\'s value is an object', () => {
+      let select = createSelect({
+        mode: 'single',
+        options: [
+          { v: { a: 1, b: 1 }, label: 1 },
+          { v: { a: 2, b: 2 }, label: 2 },
+        ],
+        value: { v: { a: 2, b: 2 }, label: 2 },
+        object: true,
+        valueProp: 'v',
+      })
+
+      expect(select.vm.isSelected({ v: { a: 2, b: 2 }, label: 2 })).toBe(true)
+      expect(select.vm.isSelected({ v: { a: 1, b: 1 }, label: 1 })).toBe(false)
     })
   })